refactor: update model template
