+++ align_anything/configs/train/text_to_text/rm.yaml
@@ -26,7 +26,7 @@ train_cfgs:
   # Batch size per device for evaluation
   per_device_eval_batch_size: 4
   # The number of gradient accumulation steps
-  gradient_accumulation_steps: 8
+  gradient_accumulation_steps: 2
   # Whether to use gradient checkpointing
   gradient_checkpointing: True
   # Initial learning rate
