+++ align_anything/datasets/text_image_to_text_image/supervised.py
@@ -83,7 +83,7 @@ class SupervisedDataset(Dataset):
         )
         if size:
             self.raw_data = self.raw_data.select(range(int(size)))
-        self.template = get_template_class(template)
+        self.template = template
 
     def preprocess(self, raw_sample: dict[str, Any]) -> SupervisedSample:
         formatted_sample = self.template.format_supervised_sample(raw_sample)
@@ -169,7 +169,7 @@ class SupervisedTokenizedDataset(Dataset):
         self.raw_data = torch.load(f"{path}/{data_files}", map_location=torch.device('cpu'))
         if size:
             self.raw_data = self.raw_data.select(range(int(size)))
-        self.template = get_template_class(template)
+        self.template = template
 
     def get_collator(self) -> Callable[[list[dict[str, torch.Tensor]]], dict[str, torch.Tensor]]:
         return SupervisedCollator(self.tokenizer.pad_token_id)
