+++ align_anything/datasets/text_to_text/preference.py
@@ -24,7 +24,6 @@ from torchvision import transforms
 from transformers.tokenization_utils import PaddingStrategy, TruncationStrategy
 
 from align_anything.utils.multi_process import get_current_device
-from align_anything.utils.template_registry import get_template_class
 from align_anything.utils.tools import right_padding
 from datasets import load_dataset
 
@@ -67,7 +66,7 @@ class PreferenceDataset(Dataset):
         assert template, f'You must set the valid template path! Here is {template}'
         self.tokenizer = tokenizer
         self.processor = processor
-        self.template = get_template_class(template)
+        self.template = template
 
         if isinstance(optional_args, str):
             optional_args = [optional_args]
@@ -89,27 +88,17 @@ class PreferenceDataset(Dataset):
         valid_indices = []
         for i, item in enumerate(self.raw_data):
             if not self.template.check_equal(item):
+                if hasattr(self.template, 'check_validation'):
+                    if not self.template.check_validation(item):
+                        continue
                 valid_indices.append(i)
         return valid_indices
 
     def preprocess(self, raw_sample: dict[str, Any]) -> PreferenceSample:
-        formatted_sample = self.template.format_preference_sample(raw_sample)
         return_dict = {}
-
-        raw_better_text = ''
-        raw_worse_text = ''
-
-        if isinstance(formatted_sample['better_text'], list):
-            raw_better_text = self.tokenizer.eos_token.join(formatted_sample['better_text'])
-            raw_worse_text = self.tokenizer.eos_token.join(formatted_sample['worse_text'])
-        elif isinstance(formatted_sample['better_text'], str):
-            raw_better_text = formatted_sample['better_text'] + self.tokenizer.eos_token
-            raw_worse_text = formatted_sample['worse_text'] + self.tokenizer.eos_token
-        else:
-            raise NotImplementedError
-        
-        return_dict['better_input_ids'] = self.tokenize(raw_better_text)
-        return_dict['worse_input_ids'] = self.tokenize(raw_worse_text)
+        formatted_better_text, formatted_worse_text, _ = self.template.format_preference_sample(raw_sample)
+        return_dict['better_input_ids'] = self.tokenize(formatted_better_text)
+        return_dict['worse_input_ids'] = self.tokenize(formatted_worse_text)
 
         return return_dict
 
@@ -163,116 +152,6 @@ class PreferenceCollator:
         return_dict['input_ids'] = right_padding(input_ids, padding_value=self.pad_token_id).to(
             current_device
         )  # size = (2 * B, L)
-        attention_mask = [
-            input_id.new_ones(input_id.size(), dtype=torch.bool) for input_id in input_ids
-        ]  # size = (2 * B, L)
-        return_dict['attention_mask'] = right_padding(attention_mask, padding_value=0).to(
-            current_device
-        )  # size = (2 * B, L)
-
-        return return_dict
-
-
-class RandomPreferenceDataset(Dataset):
-
-    def __init__(
-        self,
-        path: str,
-        template: str,
-        tokenizer: transformers.PreTrainedTokenizer,
-        processor: transformers.ProcessorMixin | transforms.Compose | None = None,
-        size: int | None = None,
-        split: str | None = None,
-        data_files: str | None = None,
-        optional_args: list | str = [],
-    ):
-        super().__init__()
-        self.tokenizer = tokenizer
-        self.processor = processor
-        self.raw_data = raw_data_duplicated = load_dataset(
-            path,
-            name=name,
-            split=split,
-            data_files=data_files,
-            *optional_args,
-            trust_remote_code=True,
-        )
-        if size:
-            self.raw_data = self.raw_data.select(range(int(size)))
-        self.template = get_template_class(template)
-
-    def preprocess(
-        self, raw_sample_1: dict[str, Any], raw_sample_2: dict[str, Any]
-    ) -> PreferenceSample:
-
-        formatted_sample_2 = self.template.format_preference_sample(raw_sample_2)['better_text']
-        assert isinstance(formatted_sample_2, str)
-        formatted_prompt_1 = self.template.format_prompt_only_sample(raw_sample_1)['text']
-        formatted_prompt_2 = self.template.format_prompt_only_sample(raw_sample_2)['text']
-        formatted_response_2 = formatted_sample_2.removeprefix(formatted_prompt_2)
-        formatted_text = formatted_prompt_1 + formatted_response_2
-        return_dict = {}
-        raw_better_text = ''
-        raw_better_text = formatted_text + self.tokenizer.eos_token
-        return_dict['better_input_ids'] = self.tokenize(raw_better_text)
-        return_dict['worse_input_ids'] = self.tokenize(raw_better_text)
-        return return_dict
-
-    def get_collator(self) -> Callable[[list[dict[str, torch.Tensor]]], dict[str, torch.Tensor]]:
-        return RandomPreferenceCollator(self.tokenizer.pad_token_id)
-
-    def tokenize(
-        self,
-        text: str,
-        add_special_tokens: bool = True,
-        padding: bool | str | PaddingStrategy = PaddingStrategy.DO_NOT_PAD,
-        truncation: bool | str | TruncationStrategy = TruncationStrategy.LONGEST_FIRST,
-        max_length: int | None = None,
-    ) -> torch.LongTensor:  # size = (L,)
-        """Tokenize a text string into a tensor representation."""
-        if max_length is None:
-            max_length = self.tokenizer.model_max_length
-
-        return self.tokenizer(
-            text,
-            add_special_tokens=add_special_tokens,
-            padding=padding,
-            max_length=max_length,
-            truncation=truncation,
-            return_tensors='pt',
-        )['input_ids'][0]
-
-    def __getitem__(self, index: int) -> dict[str, torch.Tensor]:
-        """Get a tokenized data sample by index."""
-        raw_sample = self.raw_data[index]
-        if index == 0:
-            raw_sample_change = self.raw_data[-1]
-        else:
-            raw_sample_change = self.raw_data[index - 1]
-        data = self.preprocess(raw_sample, raw_sample_change)
-        return data
-
-    def __len__(self) -> int:
-        """Get the number of samples in the dataset."""
-        return len(self.raw_data)
-
-
-class RandomPreferenceCollator:
-
-    def __init__(self, pad_token_id: int) -> None:
-        """Initialize a collator."""
-        self.pad_token_id = pad_token_id
-
-    def __call__(self, samples: list[PreferenceSample]) -> tuple[PreferenceBatch]:
-        return_dict = {}
-        current_device = get_current_device()
-
-        input_ids = [sample['better_input_ids'] for sample in samples]
-        # size = (B, L)
-        return_dict['input_ids'] = right_padding(input_ids, padding_value=self.pad_token_id).to(
-            current_device
-        )  # size = (2 * B, L)
-
         attention_mask = [
             input_id.new_ones(input_id.size(), dtype=torch.bool) for input_id in input_ids
         ]  # size = (2 * B, L)
