+++ align_anything/datasets/text_video_to_text/prompt_only.py
@@ -247,7 +247,7 @@ class PromptOnlyDataset(Dataset):
         )
         if size:
             self.raw_data = self.raw_data.select(range(int(size)))
-        self.template = get_template_class(template)
+        self.template = template
 
     
     def preprocess(self, raw_sample: dict[str, Any]) -> PromptOnlySample:
