+++ align_anything/datasets/text_video_to_text/supervised.py
@@ -247,7 +247,7 @@ class SupervisedDataset(Dataset):
         )
         if size:
             self.raw_data = self.raw_data.select(range(int(size)))
-        self.template = get_template_class(template)
+        self.template = template
