+++ align_anything/models/qwen2_audio.py
@@ -13,6 +13,7 @@
 # limitations under the License.
 # ==============================================================================
 
+import deepspeed
 import torch
 import torch.utils.checkpoint
 import torch.nn as nn
@@ -31,9 +32,17 @@ except ImportError:
 if Qwen2Audio_AVALIABLE:
     class AccustomedQwen2AudioModel(Qwen2AudioForConditionalGeneration):
         
+        def __init__(self, config: AutoConfig):
+            super().__init__(config)
+            deepspeed.zero.register_external_parameter(self, self.audio_tower.embed_positions.weight)
+
+        @property
+        def processor_available(self):
+            return True
+
         @property
         def infer_required_keys(self) -> list[str]:
-            return ['input_ids', 'attention_mask', 'input_features', 'feature_attention_mask']
+            return ['input_ids', 'attention_mask', 'input_features', 'feature_attention_mask', 'labels']
         
         def infer_batch(self, batch: dict[str, torch.Tensor]) -> dict[str, torch.Tensor]:
             """Return the dict used for model inference"""
@@ -42,6 +51,7 @@ if Qwen2Audio_AVALIABLE:
                 'attention_mask': batch['attention_mask'],
                 'input_features': batch['input_features'],
                 'feature_attention_mask': batch['feature_attention_mask'],
+                'labels': batch.get('labels'),
             }
 
     class AccustomedQwen2AudioRewardModel(Qwen2AudioPreTrainedModel):
@@ -50,12 +60,17 @@ if Qwen2Audio_AVALIABLE:
         def __init__(self, config: AutoConfig):
             super().__init__(config)
             setattr(self, self.base_model_prefix, AccustomedQwen2AudioModel(config))
+            deepspeed.zero.register_external_parameter(self.model, self.model.audio_tower.embed_positions.weight)
             self.score_head = nn.Linear(4096, 1, bias=False)
 
         @property
         def infer_required_keys(self) -> list[str]:
             return ['input_ids', 'attention_mask', 'input_features', 'feature_attention_mask']
         
+        @property
+        def processor_available(self):
+            return True
+
         def infer_batch(self, batch: dict[str, torch.Tensor]) -> dict[str, torch.Tensor]:
             """Return the dict used for model inference"""
             return {
