+++ align_anything/trainers/text_to_text/orpo.py
@@ -30,7 +30,7 @@ from transformers.integrations.deepspeed import HfDeepSpeedConfig
 
 from align_anything.datasets.text_to_text.preference import PreferenceBatch, PreferenceDataset
 from align_anything.models.pretrained_model import load_pretrained_models
-from align_anything.trainers.base import SupervisedTrainerBase
+from align_anything.trainers.text_to_text.dpo import DPOTrainer
 from align_anything.utils.multi_process import (
     get_all_reduce_mean,
     get_current_device,
@@ -48,81 +48,7 @@ from align_anything.utils.tools import (
 )
 
 
-class ORPOTrainer(SupervisedTrainerBase):
-
-    def __init__(self, cfgs, ds_cfgs) -> None:
-        """Initialize trainer."""
-        self.cfgs = cfgs
-        self.ds_train_cfgs = prepare_ds_train_cfgs(custom_cfgs=cfgs.train_cfgs, raw_ds_cfgs=ds_cfgs)
-        self.ds_eval_cfgs = prepare_ds_eval_cfgs(custom_cfgs=cfgs.train_cfgs, raw_ds_cfgs=ds_cfgs)
-        self.global_step = 0
-
-        self.init_check()
-        dist.barrier()
-        self.init_models()
-        if hasattr(self.model, 'infer_batch'):
-            self.infer_batch = self.model.infer_batch
-        if hasattr(self.model, 'infer_required_keys'):
-            self.infer_required_keys = self.model.infer_required_keys
-        dist.barrier()
-        self.init_datasets()
-        dist.barrier()
-        self.init_engines()
-        dist.barrier()
-        self.init_logger()
-
-    def init_check(self) -> None:
-        """Initial configuration checking."""
-        super().init_check()
-
-    def init_models(self) -> None:
-        """Initialize model and tokenizer."""
-        if self.ds_train_cfgs['zero_optimization']['stage'] == 3:
-            self.dstchf_train = HfDeepSpeedConfig(self.ds_train_cfgs)
-        if self.ds_eval_cfgs['zero_optimization']['stage'] == 3:
-            self.dsechf_eval = HfDeepSpeedConfig(self.ds_eval_cfgs)
-        self.bnb_cfgs = self.cfgs.bnb_cfgs
-        self.lora_cfgs = self.cfgs.lora_cfgs
-        self.model, self.tokenizer, self.processor = load_pretrained_models(
-            self.cfgs.model_cfgs.model_name_or_path,
-            model_max_length=self.cfgs.model_cfgs.model_max_length,
-            padding_side='left',
-            trust_remote_code=True,
-            bnb_cfgs=self.bnb_cfgs,
-            lora_cfgs=self.lora_cfgs,
-        )
-        self.reference_model, _, _ = load_pretrained_models(
-            self.cfgs.model_cfgs.model_name_or_path,
-            model_max_length=self.cfgs.model_cfgs.model_max_length,
-            padding_side='left',
-            trust_remote_code=True,
-            bnb_cfgs=self.bnb_cfgs,
-            lora_cfgs=self.lora_cfgs,
-        )
-
-    def init_datasets(self) -> None:
-        """Initialize training and evaluation datasets."""
-        self.train_dataloader, self.eval_dataloader = self.get_dataloaders(
-            PreferenceDataset, PreferenceDataset
-        )
-
-    def init_engines(self) -> None:
-        """Initialize DeepSpeed engines."""
-        self.init_deepspeed_engines()
-        self.reference_model, *_ = deepspeed.initialize(
-            model=self.reference_model,
-            config=self.ds_eval_cfgs,
-        )
-
-    def compute_log_probs(
-        self,
-        model: AutoModelForCausalLM,
-        batch: PreferenceBatch,
-    ) -> torch.Tensor:
-        """Compute log probabilities of given sequences."""
-        logits = model(**self.infer_batch(batch)).logits
-        input_ids = batch['input_ids']
-        return gather_log_probabilities(logits[:, :-1], input_ids[:, 1:])
+class ORPOTrainer(DPOTrainer):
 
     def loss(  # pylint: disable=too-many-locals
         self,
@@ -137,15 +63,6 @@ class ORPOTrainer(SupervisedTrainerBase):
             better_sequence_log_probs,  # size = (B, L - 1)
             worse_sequence_log_probs,  # size = (B, L - 1)
         ) = sequence_log_probs.chunk(chunks=2, dim=0)
-
-        with torch.no_grad():
-            ref_sequence_log_probs = self.compute_log_probs(  # size = (2 * B, L - 1)
-                self.reference_model.module,
-                batch,
-            )
-            ref_better_sequence_log_probs, ref_worse_sequence_log_probs = (
-                ref_sequence_log_probs.chunk(chunks=2, dim=0)
-            )
 
         losses = []
         better_sample_rewards = []
@@ -205,107 +122,6 @@ class ORPOTrainer(SupervisedTrainerBase):
             'reward_accuracy': reward_accuracy,
             'reward_margin': reward_margin,
         }
-
-    def train_step(
-        self,
-        batch: PreferenceBatch,
-    ) -> dict[str, Any]:
-        """Perform a single training step for ORPO."""
-        loss_dict = self.loss(batch=batch)
-        loss = loss_dict['loss']
-        self.model.backward(loss)
-        self.model.step()
-
-        with torch.no_grad():
-            reward = loss_dict['reward'].mean()
-            better_sample_reward = loss_dict['better_sample_reward'].mean()
-            worse_sample_reward = loss_dict['worse_sample_reward'].mean()
-            reward_accuracy = loss_dict['reward_accuracy']
-            reward_margin = loss_dict['reward_margin'].mean()
-
-            loss = get_all_reduce_mean(loss)
-            reward = get_all_reduce_mean(reward)
-            better_sample_reward = get_all_reduce_mean(better_sample_reward)
-            worse_sample_reward = get_all_reduce_mean(worse_sample_reward)
-            reward_accuracy = get_all_reduce_mean(reward_accuracy)
-            reward_margin = get_all_reduce_mean(reward_margin)
-
-        return {
-            'train/loss': loss.item(),
-            'train/reward': reward.item(),
-            'train/better_sample_reward': better_sample_reward.item(),
-            'train/worse_sample_reward': worse_sample_reward.item(),
-            'train/reward_accuracy': reward_accuracy.item(),
-            'train/reward_margin': reward_margin.item(),
-            'train/lr': self.model.optimizer.param_groups[0]['lr'],
-        }
-
-    def train(self) -> None:
-        """Train the model."""
-        self.logger.print('***** Running training *****')
-
-        progress_bar = tqdm(
-            total=self.cfgs.train_cfgs.epochs * len(self.train_dataloader),
-            desc=f'Training 1/{self.cfgs.train_cfgs.epochs} epoch',
-            position=0,
-            leave=True,
-            disable=not is_main_process(),
-        )
-
-        if self.cfgs.data_cfgs.eval_datasets:
-            self.logger.print('\n***** Evaluating at the beginning *****')
-            self.logger.log(self.eval(), step=0)
-
-        for epoch in range(int(self.cfgs.train_cfgs.epochs)):
-            self.model.train()
-
-            for batch in self.train_dataloader:
-                info = self.train_step(batch)
-                torch.cuda.empty_cache()
-
-                self.global_step += 1
-                progress_bar.set_description(
-                    f'Training {epoch + 1}/{self.cfgs.train_cfgs.epochs} epoch '
-                    f'(loss {info["train/loss"]:.4f})',
-                )
-                progress_bar.update(1)
-
-                info['train/epoch'] = self.global_step / len(self.train_dataloader)
-                self.logger.log(info, step=self.global_step)
-
-                if self.global_step % self.cfgs.logger_cfgs.save_interval == 0:
-                    self.logger.print(f'Saving checkpoint at step {self.global_step} ...')
-                    self.save(tag=self.global_step)
-                    self.logger.print('Checkpoint saved.')
-
-                if (
-                    self.cfgs.data_cfgs.eval_datasets
-                    and self.cfgs.train_cfgs.eval_strategy == 'steps'
-                    and self.global_step % self.cfgs.train_cfgs.eval_interval == 0
-                ):
-                    self.logger.print(f'\n***** Evaluating at step {self.global_step} *****')
-                    self.logger.log(self.eval(), step=self.global_step)
-
-            if self.cfgs.data_cfgs.eval_datasets and self.cfgs.train_cfgs.eval_strategy == 'epoch':
-                self.logger.print(
-                    f'\n***** Evaluating at epoch {epoch + 1}/{self.cfgs.train_cfgs.epochs} *****',
-                )
-                self.logger.log(self.eval(), step=self.global_step)
-
-            self.model.tput_timer.update_epoch_count()
-
-    @torch.no_grad()
-    def eval(self) -> dict[str, Any]:
-        """Evaluate the model on the evaluation dataset."""
-        return {}
-
-    def save(
-        self,
-        model: deepspeed.DeepSpeedEngine | None = None,
-        tag: int | None = None,
-    ) -> None:
-        """Save model and tokenizer in Hugging Face format."""
-        self.save_transformers(model=model, tag=tag)
 
 
 def main():
