+++ docs/source/training/text_image_to_text.rst
@@ -2,26 +2,21 @@ Text Image to Text Scripts
 ==========================
 
 We provide examples of various scripts for finetuning based on the
-`LLaVA <https://huggingface.co/llava-hf>`__ series models as follows. You can execute these commands
+`PKU-Alignment/Align-Anything-TI2T-Instruction-100K <https://huggingface.co/datasets/PKU-Alignment/Align-Anything-TI2T-Instruction-100K>`__ and `PKU-Alignment/align-anything-400k <https://huggingface.co/datasets/PKU-Alignment/align-anything-400k>`__ series datasets as follows. You can execute these commands
 under the ``./scripts`` to start the corresponding training.
 
 Supervised Fine-Tuning
 ----------------------
 
-.. warning::
-
-    The `LLaVA-Instruct-150K <https://huggingface.co/datasets/liuhaotian/LLaVA-Instruct-150K>`__ dataset uses remote links to access images on `coco 2017 <https://cocodataset.org/#home>`__. However, due to the server hosting these images occasionally stopping its response, this can lead to unstable training. We strongly recommend that you download the `coco2017 dataset <http://images.cocodataset.org/zips/train2017.zip>`__ locally here, then update the image paths in ``llava_instruct_150k.json`` to your local paths, and then use the ``Llava_Local`` template.
-
 .. code:: bash
 
    MODEL_NAME_OR_PATH="llava-hf/llava-1.5-7b-hf" # model path
 
-   TRAIN_DATASETS="liuhaotian/LLaVA-Instruct-150K" # dataset path
-   TRAIN_TEMPLATE="Llava" # dataset template
+   TRAIN_DATASETS="PKU-Alignment/Align-Anything-TI2T-Instruction-100K" # dataset path
+   TRAIN_TEMPLATE="AA_TI2T" # dataset template
    TRAIN_SPLIT="train" # split the dataset
-   TRAIN_DATA_FILES="llava_instruct_150k.json" # specify the training data files
 
-   OUTPUT_DIR="../output/sft" # output dir
+   OUTPUT_DIR="../outputs/sft_llava" # output dir
 
    # For wandb online logging
    export WANDB_API_KEY=""
@@ -35,10 +30,11 @@ Supervised Fine-Tuning
         --module align_anything.trainers.text_image_to_text.sft \
         --model_name_or_path ${MODEL_NAME_OR_PATH} \
         --train_datasets ${TRAIN_DATASETS} \
-        --output_dir ${OUTPUT_DIR} \
         --train_template ${TRAIN_TEMPLATE} \
         --train_split ${TRAIN_SPLIT} \
-        --train_data_files ${TRAIN_DATA_FILES}
+        --output_dir ${OUTPUT_DIR} \
+        --save_interval 1200 \
+        --epochs 3
 
 Reward Model Training
 ---------------------
@@ -47,11 +43,17 @@ Reward Model Training
 
    MODEL_NAME_OR_PATH="llava-hf/llava-1.5-7b-hf" # model path
 
-   TRAIN_DATASETS="sqrti/SPA-VL" # dataset path
-   TRAIN_TEMPLATE="SPA_VL" # dataset template
+   TRAIN_DATASETS="PKU-Alignment/align-anything-400k" # dataset path
+   TRAIN_TEMPLATE="AA_TI2T" # dataset template
+   TRAIN_NAME="text-image-to-text" # dataset name
    TRAIN_SPLIT="train" # split the dataset
 
-   OUTPUT_DIR="../output/rm" # output dir
+   EVAL_DATASETS="PKU-Alignment/align-anything-400k" # dataset path
+   EVAL_TEMPLATE="AA_TI2T" # dataset template
+   EVAL_NAME="text-image-to-text" # dataset name
+   EVAL_SPLIT="val" # split the dataset
+
+   OUTPUT_DIR="../outputs/rm_llava" # output dir
 
    # For wandb online logging
    export WANDB_API_KEY=""
@@ -66,8 +68,16 @@ Reward Model Training
         --model_name_or_path ${MODEL_NAME_OR_PATH} \
         --train_datasets ${TRAIN_DATASETS} \
         --train_template ${TRAIN_TEMPLATE} \
+        --train_name ${TRAIN_NAME} \
         --train_split ${TRAIN_SPLIT} \
-        --output_dir ${OUTPUT_DIR}
+        --eval_datasets ${EVAL_DATASETS} \
+        --eval_template ${EVAL_TEMPLATE} \
+        --eval_name ${EVAL_NAME} \
+        --eval_split ${EVAL_SPLIT} \
+        --output_dir ${OUTPUT_DIR} \
+        --save_interval 1200 \
+        --epochs 5
+
 
 DPO Training
 ------------
@@ -76,11 +86,12 @@ DPO Training
 
    MODEL_NAME_OR_PATH="llava-hf/llava-1.5-7b-hf" # model path
 
-   TRAIN_DATASETS="sqrti/SPA-VL" # dataset path
-   TRAIN_TEMPLATE="SPA_VL" # dataset template
+   TRAIN_DATASETS="PKU-Alignment/align-anything-400k" # dataset path
+   TRAIN_TEMPLATE="AA_TI2T" # dataset template
+   TRAIN_NAME="text-image-to-text" # dataset name
    TRAIN_SPLIT="train" # split the dataset
 
-   OUTPUT_DIR="../output/dpo" # output dir
+   OUTPUT_DIR="../outputs/dpo_llava" # output dir
 
    # For wandb online logging
    export WANDB_API_KEY=""
@@ -95,8 +106,11 @@ DPO Training
         --model_name_or_path ${MODEL_NAME_OR_PATH} \
         --train_datasets ${TRAIN_DATASETS} \
         --train_template ${TRAIN_TEMPLATE} \
+        --train_name ${TRAIN_NAME} \
         --train_split ${TRAIN_SPLIT} \
-        --output_dir ${OUTPUT_DIR}
+        --output_dir ${OUTPUT_DIR} \
+        --save_interval 1200 \
+        --epochs 5
 
 .. warning::
 
@@ -107,29 +121,40 @@ PPO Training
 
 .. code:: bash
 
-   ACTOR_MODEL_NAME_OR_PATH="llava-hf/llava-1.5-7b-hf" # actor model path
-   CRITIC_MODEL_NAME_OR_PATH="PATH_TO_YOUR_REWARD_MODEL" # critic model path
-   REWARD_MODEL_NAME_OR_PATH="PATH_TO_YOUR_REWARD_MODEL" # reward model path
-   
-   TRAIN_DATASETS="sqrti/SPA-VL" # rlhf dataset path
-   TRAIN_TEMPLATE="SPA_VL" # rlhf dataset template
-   TRAIN_SPLIT="train" # split the rlhf dataset
+    ACTOR_MODEL_NAME_OR_PATH="llava-hf/llava-1.5-7b-hf" # model path
+    REWARD_MODEL_NAME_OR_PATH="../outputs/rm_llava" # model path
+    CRITIC_MODEL_NAME_OR_PATH="../outputs/rm_llava" # model path
 
-   OUTPUT_DIR="../output/ppo" # output dir
-   # For wandb online logging
-   export WANDB_API_KEY=""
+    TRAIN_DATASETS="PKU-Alignment/align-anything-400k" # dataset path
+    TRAIN_TEMPLATE="AA_TI2T" # dataset template
+    TRAIN_NAME="text-image-to-text" # dataset name
+    TRAIN_SPLIT="train" # split the dataset
 
-   # Source the setup script
-   source ./setup.sh
+    PTX_DATASETS="PKU-Alignment/Align-Anything-TI2T-Instruction-100K"
+    PTX_TEMPLATE="AA_TI2T"
+    PTX_SPLIT="train"
 
-   # Execute deepspeed command
-   deepspeed \
-     --master_port ${MASTER_PORT} \
-     --module align_anything.trainers.text_image_to_text.ppo \
-     --actor_model_name_or_path ${ACTOR_MODEL_NAME_OR_PATH} \
-     --reward_model_name_or_path ${REWARD_MODEL_NAME_OR_PATH} \
-     --reward_critic_model_name_or_path ${CRITIC_MODEL_NAME_OR_PATH} \
-     --train_datasets ${TRAIN_DATASETS} \
-     --train_split ${TRAIN_SPLIT} \
-     --train_template ${TRAIN_TEMPLATE} \
-     --output_dir ${OUTPUT_DIR}
+    OUTPUT_DIR="../outputs/ppo_ti2t" # output dir
+    # For wandb online logging
+    export WANDB_API_KEY=""
+
+    # Source the setup script
+    source ./setup.sh
+
+    # Execute deepspeed command
+    deepspeed \
+        --master_port ${MASTER_PORT} \
+        --module align_anything.trainers.text_image_to_text.ppo \
+        --actor_model_name_or_path ${ACTOR_MODEL_NAME_OR_PATH} \
+        --reward_model_name_or_path ${REWARD_MODEL_NAME_OR_PATH} \
+        --reward_critic_model_name_or_path ${CRITIC_MODEL_NAME_OR_PATH} \
+        --train_datasets ${TRAIN_DATASETS} \
+        --train_template ${TRAIN_TEMPLATE} \
+        --train_split ${TRAIN_SPLIT} \
+        --train_name ${TRAIN_NAME} \
+        --ptx_datasets ${PTX_DATASETS} \
+        --ptx_template ${PTX_TEMPLATE} \
+        --ptx_split ${PTX_SPLIT} \
+        --output_dir ${OUTPUT_DIR} \
+        --save_interval 1200 \
+        --epochs 5
